mapper: Implement subtree-remove command
The command will take in a argument in the form of NAMESPACE:INTERFACE
and use the ObjectMapper GetSubtreePaths method to see if there are any
dbus object that exposes the target interface under the target
namespace.

It will make an initial call to check the starting state and then will
make a GetSubTreePaths call everytime there is an InterfacesRemoved
signal. The interface is removed if the subptree paths returned is empty
or if we hit an ResourceNotFound error. We can hit this error when the
namespace we specified does not exists anymore or if ObjectMapper
service is down. In both case, it is considered to be removed.

Signed-off-by: Willy Tu <wltu@google.com>